app/requirements.txt
python-aemet==0.4.1
plotly==5.18.0
dash==2.14.2
pandas==2.1.4
dash-bootstrap-components==1.5.0
gunicorn==21.2.0
Flask==3.0.0